+++ changes.md
@@ -2,6 +2,8 @@
 
 * enqueue rotate for multielevation
 * remove 'other' group from geojsonmarker groups if empty
+* leaflet-elevation 2.4.0 with linear gradient
+* leaflet-elevation waypoint labels rotate
 
 Please note the [install instructions](https://leafext.de/en/doku/about/versions/).